+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - added support for spectra in CSV files (row-wise)
+- added support for spectra in ARFF files (row-wise)
 
 
 0.0.2 (2025-06-18)